removing build folder
