+++ tests/test_latex/test_latex_build.tex
@@ -1,0 +1,110 @@
+\begin{document}
+
+\pagestyle{empty}
+\sphinxmaketitle
+\pagestyle{plain}
+\sphinxtableofcontents
+\pagestyle{normal}
+\phantomsection\label{\detokenize{index::doc}}
+
+
+
+\chapter{content 2}
+\label{\detokenize{algorithm/_algo_nonumber:content-2}}\label{\detokenize{algorithm/_algo_nonumber::doc}}\label{algorithm/_algo_nonumber:algorithm-0}
+\begin{sphinxadmonition}{note}{Algorithm (Test algorithm directive)}
+
+
+
+\sphinxAtStartPar
+Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
+\end{sphinxadmonition}
+
+
+\chapter{content 1}
+\label{\detokenize{algorithm/_algo_labeled_titled_with_classname:content-1}}\label{\detokenize{algorithm/_algo_labeled_titled_with_classname::doc}}\label{algorithm/_algo_labeled_titled_with_classname:test-algo-label}
+\begin{sphinxadmonition}{note}{Algorithm 1 (Test algorithm directive)}
+
+
+
+\sphinxAtStartPar
+Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
+\end{sphinxadmonition}
+
+
+\chapter{content 3}
+\label{\detokenize{algorithm/_algo_numbered_reference:content-3}}\label{\detokenize{algorithm/_algo_numbered_reference::doc}}
+\sphinxAtStartPar
+referencing with number {\hyperref[\detokenize{algorithm/_algo_labeled_titled_with_classname:test-algo-label}]{\sphinxcrossref{Algorithm 1}}}.
+
+
+\chapter{content 4}
+\label{\detokenize{algorithm/_algo_text_reference:content-4}}\label{\detokenize{algorithm/_algo_text_reference::doc}}
+\sphinxAtStartPar
+referencing with text: {\hyperref[\detokenize{algorithm/_algo_labeled_titled_with_classname:test-algo-label}]{\sphinxcrossref{text}}}.
+
+
+\chapter{content 1}
+\label{\detokenize{proof/_proof_with_classname:content-1}}\label{\detokenize{proof/_proof_with_classname::doc}}
+\begin{sphinxadmonition}{note}
+\sphinxAtStartPar
+Proof. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
+\end{sphinxadmonition}
+
+
+\chapter{content 4}
+\label{\detokenize{proof/_proof_with_labeled_math:content-4}}\label{\detokenize{proof/_proof_with_labeled_math::doc}}
+\begin{sphinxadmonition}{note}
+\sphinxAtStartPar
+Proof. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut
+\begin{equation}\label{equation:proof/_proof_with_labeled_math:label1}
+\begin{split}P_t(x, y) = \mathbb 1\{x = y\} + t Q(x, y) + o(t)\end{split}
+\end{equation}\end{sphinxadmonition}
+
+
+\chapter{content 2}
+\label{\detokenize{proof/_proof_no_classname:content-2}}\label{\detokenize{proof/_proof_no_classname::doc}}
+\begin{sphinxadmonition}{note}
+\sphinxAtStartPar
+Proof. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
+Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
+
+\sphinxAtStartPar
+Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
+\end{sphinxadmonition}
+
+
+\chapter{content 3}
+\label{\detokenize{proof/_proof_with_unlabeled_math:content-3}}\label{\detokenize{proof/_proof_with_unlabeled_math::doc}}
+\begin{sphinxadmonition}{note}
+\sphinxAtStartPar
+Proof. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut
+\begin{equation*}
+\begin{split}P_t(x, y) = \mathbb 1\{x = y\} + t Q(x, y) + o(t)\end{split}
+\end{equation*}\end{sphinxadmonition}
+
+
+\chapter{content 5}
+\label{\detokenize{proof/_proof_with_argument_content:content-5}}\label{\detokenize{proof/_proof_with_argument_content::doc}}
+\begin{sphinxadmonition}{note}
+\sphinxAtStartPar
+Proof. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
+
+\sphinxAtStartPar
+Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
+Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
+\end{sphinxadmonition}
+
+
+\renewcommand{\indexname}{Proof Index}
+\begin{sphinxtheindex}
+\let\bigletter\sphinxstyleindexlettergroup
+\bigletter{algorithm\sphinxhyphen{}0}
+\item\relax\sphinxstyleindexentry{algorithm\sphinxhyphen{}0}\sphinxstyleindexextra{algorithm/\_algo\_nonumber}\sphinxstyleindexpageref{algorithm/_algo_nonumber:\detokenize{algorithm-0}}
+\indexspace
+\bigletter{test\sphinxhyphen{}algo\sphinxhyphen{}label}
+\item\relax\sphinxstyleindexentry{test\sphinxhyphen{}algo\sphinxhyphen{}label}\sphinxstyleindexextra{algorithm/\_algo\_labeled\_titled\_with\_classname}\sphinxstyleindexpageref{algorithm/_algo_labeled_titled_with_classname:\detokenize{test-algo-label}}
+\end{sphinxtheindex}
+
+\renewcommand{\indexname}{Index}
+\printindex
+\end{document}
